Refactoring: tokens
Move token generation/decoding to single file with associated middleware file.
