+++ README.md
@@ -58,8 +58,8 @@ TK
 | File                          | Test files                             | Coverage |
 |-------------------------------|----------------------------------------|----------|
 |                               |db_connection_test.py                   |N/A       |
-|backend/controllers/users.py   |user_signup_test.py, user_login_test.py |110/195   |
 |backend/utility/db_handlers.py |db_handler_test.py                      |10/10     |
+|backend/controllers/users.py   |user_signup_test.py, user_login_test.py |110/195   |
 |           |           |           |
 |           |           |           |
 |           |           |           |
