+++ backend/app/tests.py
@@ -6,8 +6,6 @@ import unittest
 from django.test import TestCase
 from django.test import TestCase, RequestFactory
 
-
-
 class TestSignUpUser(TestCase):
     def test_signup_user_success(self):
